Adding abstract CommonLoadBalancer
