+++ core/lean/build.gradle
@@ -21,11 +21,11 @@ apply plugin: 'eclipse'
 apply plugin: 'maven'
 apply plugin: 'org.scoverage'
 
-ext.dockerImageName = 'controller'
+ext.dockerImageName = 'lean'
 apply from: '../../gradle/docker.gradle'
 distDocker.dependsOn ':common:scala:distDocker', 'distTar'
 
-project.archivesBaseName = "openwhisk-controller"
+project.archivesBaseName = "openwhisk-lean"
 
 ext.coverageJars = [
     "${buildDir}/libs/${project.archivesBaseName}-$version-scoverage.jar",
@@ -43,7 +43,7 @@ dependencies {
     compile 'com.lightbend.akka.discovery:akka-discovery-kubernetes-api_2.12:0.11.0'
     compile 'com.lightbend.akka.discovery:akka-discovery-marathon-api_2.12:0.11.0'
     compile project(':common:scala')
-    compile project(':core:invoker')
+    compile project(':core:controller')
     scoverage gradle.scoverage.deps
 }
