+++ settings.gradle
@@ -19,6 +19,7 @@ include 'common:scala'
 
 include 'core:controller'
 include 'core:invoker'
+include 'core:lean'
 
 include 'tests'
 include 'tests:performance:gatling_tests'
